+++ src/SbeCodeGenerator/Generators/Fields/ArrayFieldDefinition.cs
@@ -6,7 +6,7 @@ namespace SbeSourceGenerator
     {
         public void AppendFileContent(StringBuilder sb, int tabs = 0)
         {
-            sb.AppendSummary(Description, tabs, nameof(ArrayFieldDefinition));
+            sb.AppendSummary(Description, tabs);
             sb.AppendTabs(tabs).Append("public ReadOnlySpan<").Append(PrimitiveType).Append("> ").Append(Name).AppendLine(";");
         }
     }