+++ src/SbeCodeGenerator/Generators/Fields/CompositeRefFieldDefinition.cs
@@ -11,7 +11,7 @@ namespace SbeSourceGenerator
         public void AppendFileContent(StringBuilder sb, int tabs = 0)
         {
             string fieldName = Name.FirstCharToLower();
-            sb.AppendSummary(Description, tabs, nameof(CompositeRefFieldDefinition));
+            sb.AppendSummary(Description, tabs);
             sb.AppendTabs(tabs).Append("private ").Append(TypeName).Append(" ").Append(fieldName).AppendLine(";");
             sb.AppendTabs(tabs).Append("public ").Append(TypeName).Append(" ").Append(Name)
                 .Append(" { readonly get => ").Append(fieldName)