+++ src/SbeCodeGenerator/Generators/Fields/ConstantTypeFieldDefinition.cs
@@ -8,7 +8,7 @@ namespace SbeSourceGenerator
         public int Length => 0;
         public void AppendFileContent(StringBuilder sb, int tabs = 0)
         {
-            sb.AppendSummary(Description, tabs, nameof(ConstantTypeFieldDefinition));
+            sb.AppendSummary(Description, tabs);
             if (Value == "")
                 sb.AppendTabs(tabs).Append("public const ").Append(PrimitiveType).Append(" ").Append(Name).Append(" = (").Append(PrimitiveType).Append(")").Append(ValueRef).AppendLine(";");
             else